EcommerceProject/src/main/java/com/ecommerce/pageobjects/AccountCreationPage.java
package com.ecommerce.pageobjects;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.ecommerce.actiondriver.Action;
import com.ecommerce.base.BaseClass;

public class AccountCreationPage extends BaseClass{

	
	
	@FindBy(xpath= "//h1[text()='Create an account']")
	WebElement formTitle;
	
	public AccountCreationPage() {
		PageFactory.initElements(getDriver(), this);
	}
	
	public boolean validateAccountCreatePage() throws Throwable  {
		return Action.isDisplayed(getDriver(), formTitle);
	}
	
}
